AllMost Complete Project

+++ APP/server/Controllers/RoomItems.js
@@ -7,6 +7,9 @@ const { removeFileFromCloudinary } = require("../Utilities/FileRemover");
 const { deleteFoldersRecursively, isUserActionAllowed } = require("./Rooms");
 const Permissions = require("../Models/Permissions");
 const ChatMessage = require("../Models/ChatMessage");
+const { mapMediaTypeToCloudinaryResource, getMediaType } = require("../Data/extensionsData");
+const Room = require("../Models/Room");
+const { UploadFileToCloudinary } = require("../Utilities/FileUploader");
 
 
 
@@ -15,7 +18,8 @@ const ChatMessage = require("../Models/ChatMessage");
 // delete a folder
 exports.deleteFolder = async(req,res) => {
     try {
-        const {folderId,prevFolderId,roomId} = req.body;
+        const {folderId,prevFolderId,roomId,softDelete,softDeleteVal} = req.body;
+        // console.log(req.body);
         const user = req.user;
 
         if(!user || !folderId || !prevFolderId){
@@ -32,7 +36,7 @@ exports.deleteFolder = async(req,res) => {
         const folder = await Folder.findById(folderId);
         const room = await RoomModel.findById(roomId);
 
-        if(!reqUser || !prevFolder || !folder || !room){
+        if(!reqUser || !prevFolder || !folder || !room || folder.isRoot){
             return res.status(404).json(
                 {
                     success:false,
@@ -55,19 +59,21 @@ exports.deleteFolder = async(req,res) => {
             )
         }
 
-        // remove the folder from prev folder
-        await Folder.findByIdAndUpdate(prevFolder._id,
-            {
-                $pull:{
-                    Folders:folder._id,
+        if(!softDelete){
+            // remove the folder from prev folder
+            await Folder.findByIdAndUpdate(prevFolder._id,
+                {
+                    $pull:{
+                        Folders:folder._id,
+                    }
+                },{
+                    new:true,
                 }
-            },{
-                new:true,
-            }
-        );
+            );
+        }
 
         //delete the folder recursively
-        await deleteFoldersRecursively(folder._id);
+        await deleteFoldersRecursively(folder._id,softDelete,softDeleteVal);
 
         return res.status(200).json(
             {
@@ -91,7 +97,7 @@ exports.deleteFolder = async(req,res) => {
 // delete file
 exports.deleteFile = async(req,res) => {
     try {
-        const {fileId,roomId} = req.body;
+        const {fileId,roomId,softDelete,softDeleteVal} = req.body;
         const reqUser = req.user;
         const room = await RoomModel.findById(roomId);
 
@@ -120,16 +126,25 @@ exports.deleteFile = async(req,res) => {
         // if deletion allowed 
 
         // first pull this out of te folder
-        await Folder.findByIdAndUpdate(file.Folder,
-            {
-                $pull:{
-                    Files:file._id,
+        if(!softDelete){
+            await Folder.findByIdAndUpdate(file.Folder,
+                {
+                    $pull:{
+                        Files:file._id,
+                    }
                 }
-            }
-        );
+            );
+        }
 
         // now nothing just delete this file
-        await File.findByIdAndDelete(file._id);
+        if(softDelete){
+            await File.findByIdAndUpdate(file._id,{
+                isDeleted:softDeleteVal,
+            });
+        }
+        else{
+            await File.findByIdAndDelete(file._id);
+        }
 
         // all success
         return res.status(200).json(
@@ -153,7 +168,9 @@ exports.deleteFile = async(req,res) => {
 // delete a media
 exports.deleteMedia = async(req,res) => {
     try {
-        const {mediaId,roomId} = req.body;
+        const {mediaId,roomId,softDelete,softDeleteVal} = req.body;
+
+        console.log(req.body);
         const reqUser = req.user;
         const room = await RoomModel.findById(roomId);
 
@@ -181,20 +198,31 @@ exports.deleteMedia = async(req,res) => {
 
         // if deletion allowed 
 
-        // first pull this out of te folder
-        await Folder.findByIdAndUpdate(media.Folder,
-            {
-                $pull:{
-                    Files:media._id,
+        if(!softDelete){
+            // first pull this out of te folder
+            await Folder.findByIdAndUpdate(media.Folder,
+                {
+                    $pull:{
+                        Files:media._id,
+                    }
                 }
-            }
-        );
+            );
 
-        //delete the media fro cloudinary
-        await removeFileFromCloudinary(media.url);
+            //delete the media fro cloudinary
+            await removeFileFromCloudinary(media.url,mapMediaTypeToCloudinaryResource(media.mediaType));
+        }
 
-        // now nothing just delete this Media
-        await Media.findByIdAndDelete(media._id);
+        
+
+       if(softDelete){
+            await Media.findByIdAndUpdate(media._id,{
+                isDeleted:softDeleteVal,
+            })
+       }
+       else{
+            // now nothing just delete this Media
+            await Media.findByIdAndDelete(media._id);
+        }
 
         // all success
         return res.status(200).json(
@@ -226,6 +254,7 @@ exports.renameFile = async (req, res) => {
     const room = await RoomModel.findById(roomId);
 
     if(!reqUser || !reqFile || !room || !newName){
+        console.log(reqUser,reqFile,room,newName);
         return res.status(404).json({
           success: false,
           message: "user or file Not Found"
@@ -235,12 +264,12 @@ exports.renameFile = async (req, res) => {
     const permisiions = await Permissions.findOne(
         {
             Room:room._id,
-            User:room._id,
+            User:reqUser._id,
 
         }
     );
-
-    if(reqUser._id!==room.owner && !permisiions.write){
+    console.log(permisiions);
+    if( (!reqUser._id.equals(room.owner)) && ((!permisiions) || (!permisiions.write))){
         return res.status(401).json({
           success: false,
           message: "User Action Not Allowed",
@@ -293,11 +322,11 @@ exports.renamefolder = async (req, res) => {
     const permisiions = await Permissions.findOne(
         {
             Room:room._id,
-            User:room._id,
+            User:reqUser._id,
         }
     );
 
-    if(reqUser._id!==room.owner && !permisiions.write){
+    if( (!reqUser._id.equals(room.owner)) && ((!permisiions) || (!permisiions.write))){
         return res.status(401).json({
           success: false,
           message: "User Action Not Allowed",
@@ -351,11 +380,11 @@ exports.renameMedia = async (req, res) => {
     const permisiions = await Permissions.findOne(
         {
             Room:room._id,
-            User:room._id,
+            User:reqUser._id,
         }
     );
 
-    if(reqUser._id!==room.owner && !permisiions.write){
+    if( (!reqUser._id.equals(room.owner)) && ((!permisiions) || (!permisiions.write))){
         return res.status(401).json({
           success: false,
           message: "User Action Not Allowed",
@@ -389,12 +418,13 @@ exports.renameMedia = async (req, res) => {
 };
 
 // get the folder details 
-exports.getFolderDetails = async(req,res) => {
+exports.getItemDetails = async(req,res) => {
     try {
-        const {folderId,roomId} = req.body;
+        const {itemId,roomId,type} = req.body;
         const reqUser = req.user;
-
-        if(!folderId || !roomId || !reqUser){
+        const types = ['file','folder','media'];
+        if(!itemId || !roomId || !reqUser || !types.includes(type)){
+            console.log(itemId,roomId,reqUser,type);
             return res.status(400).json(
                 {
                     success:false,
@@ -405,14 +435,27 @@ exports.getFolderDetails = async(req,res) => {
 
         const room = await RoomModel.findById(roomId);
         // console.log(room);
-        const folder = await Folder.findById(folderId)
-        .populate('Files')
-        .populate('Medias')
-        .exec();
+        let item = null;
+        if(type===types[0]){
+            item = await File.findById(itemId);
+        }
+        else if(type===types[1]){
+            item = await Folder.findById(itemId);
+        }
+        else{
+            item = await Media.findById(itemId);
+        }
+
+        if(!item){
+            return res.status(404).json({
+              success: false,
+              message: "Item Not found"
+            });
+        }
 
         // console.log(folder);
 
-        const isReadAloowed = await isUserActionAllowed(room,reqUser,'read',folder);
+        const isReadAloowed = await isUserActionAllowed(room,reqUser,'read',item);
 
         if(!isReadAloowed && !room.owner===reqUser._id){
             return res.status(401).json(
@@ -426,17 +469,17 @@ exports.getFolderDetails = async(req,res) => {
         return res.status(200).json(
             {
                 success:true,
-                message:"Folder Info Fetched",
-                folder:folder,
+                message:"Item Info Fetched",
+                item:item,
             }
         );
     } catch (error) {
         console.error(error);
-        console.log('Some Problem Occurred in Getting Folder Info');
+        console.log('Some Problem Occurred in Getting Item Info');
         return res.status(500).json(
             {
                 success:false,
-                message:"Some Problem Occured in Folder Info",
+                message:"Some Problem Occured in Item Info",
             }
         )
     }
@@ -486,3 +529,131 @@ exports.getRoomChats = async (req, res) => {
     });
   }
 };
+
+exports.createAItem = async (req, res) => {
+  try {
+    const user = req.user;
+    const {parentFolder,type,roomId,name,language,extension} = req.body;
+    console.log(req.body);
+    const newFile = req.files?.newFile;
+    const types = ['file','folder','media'];
+
+    if(!(user._id) || !parentFolder || !type || !types.includes(type)){
+        return res.status(404).json({
+          success: false,
+          message: "Not Found Enough Details"
+        });
+    }
+    let item = null;
+    const room = await Room.findById(roomId);
+    const isAllowed = isUserActionAllowed(room,user,'write');
+
+    if(!isAllowed){
+        return res.status(401).json({
+          success: false,
+          message: "Unthorised access"
+        });
+    }
+
+    if(type===types[0]){
+        item = await File.create({
+            Folder:parentFolder,
+            owner:user._id,
+            name,
+            language,
+            Room:room._id,
+        });
+        await Folder.findByIdAndUpdate(parentFolder,{
+            $push:{
+                Files:item._id,
+            }
+        })
+    }
+    else if(type===types[1]){
+        item = await Folder.create({
+            isRoot:false,
+            name,
+            owner:room.user_id,
+            parentFolder,
+            Room:room._id,
+        });
+        await Folder.findByIdAndUpdate(parentFolder,{
+            $push:{
+                Folders:item._id,
+            }
+        })
+    }
+    else{
+        const uploadDetails = await UploadFileToCloudinary(newFile,'CodeSyncRoomData-'+room._id);
+        item = await Media.create({
+            Folder:parentFolder,
+            mediaType:getMediaType(extension),
+            name,
+            owner:user._id,
+            Room:room._id,
+            url:uploadDetails.secure_url,
+        })
+        await Folder.findByIdAndUpdate(parentFolder,{
+            $push:{
+                Medias:item._id,
+            }
+        })
+    }
+
+    return res.status(200).json({
+      success: true,
+      message: "Item Added To Room",
+      item,
+    });
+
+  } catch (error) {
+    console.error(error);
+    console.log('Some Internal Problem in createAItem');
+    return res.status(500).json({
+      success: false,
+      message: 'Some Internal Problem in creating A Item'
+    });
+  }
+};
+
+exports.saveAFile = async(req,res) => {
+    try {
+        const user = req.user;
+        const {roomId,fileId,newContent} = req.body;
+
+        if(!user._id || !roomId || !fileId){
+            return res.status(44).json({
+              success: false,
+              message: "User Or File Not Found"
+            });
+        }
+
+        const room = await Room.findById(roomId);
+        const isAllowed = await isUserActionAllowed(room,user,'write');
+        if(!isAllowed){
+            return res.status(401).json({
+              success: false,
+              message: "User Action Not Allowed"
+            });
+        }
+
+        const newFile = await File.findByIdAndUpdate(fileId,{
+            content:newContent,
+        },{new:true});
+
+        return res.status(200).json(
+            {
+                success:true,
+                message:"File Saving Successfull",
+                newFile,
+            }
+        )
+    } catch (error) {
+        console.error(error);
+        console.log('Some Internal Error in Saving The file');
+        return res.status(500).json({
+          success: false,
+          message: "Some Internal Error in Saving The file"
+        });
+    }
+}
